fix(drivers): fix ranker driver
